Internal: move timestamp component to SourceInfo panel
This is the only consumer of this component. Clean up the component
to remove the unused bits.
